Support fluentd 1.2.6 using rh-ruby25 scl base image
Use the rh-ruby25 scl base image for fluentd.  This allows us
to use the latest fluentd 1.2.6 and the latest components.

This means we do not have to carry the patches for prometheus
and concat.  We can also use the log rotation built-in to
fluentd 1.x.  There is no longer a separate secure_forward
plugin - the built-in forward plugin has TLS and secret_key
support.  The latest prometheus has TLS support.

Some of our internal plugins like parser_json_or_crio had to
be ported to use the fluentd 1.x internal api.

Another big change is that there is a new format for the
fluent-plugin-system pos file - It is in JSON format and
must have a .json suffix.  The fluentd run.sh attempts to
convert the old file to the new format file.

The config file format is a little stricter - it is
apparently now an error to have a @type copy plugin
with no <store> sections.  You also cannot have a
rewrite_tag plugin with an empty pattern.

Also cleaned up the fluentd-forward test - it needed to
use the genid plugin when sending to two outputs.

+++ test/mux-client-mode.sh
@@ -33,16 +33,12 @@ cleanup() {
     if [ -n "${muxpod:-}" ] ; then
         get_mux_pod_log $muxpod > $ARTIFACT_DIR/mux-client-mode-mux-pod.log 2>&1
     fi
-    if [ -n "${saveds:-}" ] ; then
-        if [ -f "${saveds:-}" ]; then
-            oc replace --force -f $saveds 2>&1 | artifact_out
-            rm -f $saveds
-        fi
+    stop_fluentd "${fpod:-}" $FLUENTD_WAIT_TIME 2>&1 | artifact_out
+    if [ -n "${saveds:-}" -a -f "${saveds:-}" ]; then
+        oc replace --force -f $saveds 2>&1 | artifact_out
+        rm -f $saveds
     fi
-    os::cmd::expect_success flush_fluentd_pos_files
-    sudo rm -f /var/log/fluentd/fluentd.log
-    oc label node --all logging-infra-fluentd=true 2>&1 | artifact_out
-    os::cmd::try_until_text "oc get pods -l component=fluentd" "^logging-fluentd-.* Running "
+    start_fluentd true 2>&1 | artifact_out
     # this will call declare_test_end, suite_end, etc.
     os::test::junit::reconcile_output
     exit $return_code
@@ -62,26 +58,18 @@ fpod=$( get_running_pod fluentd )
 muxpod=$( get_running_pod mux )
 
 os::log::info configure fluentd to use MUX_CLIENT_MODE=minimal - verify logs get through
-oc label node --all logging-infra-fluentd- 2>&1 | artifact_out
-os::cmd::try_until_text "oc get daemonset logging-fluentd -o jsonpath='{ .status.numberReady }'" "0" $FLUENTD_WAIT_TIME
+stop_fluentd "$fpod" $FLUENTD_WAIT_TIME 2>&1 | artifact_out
 oc set env daemonset/logging-fluentd MUX_CLIENT_MODE=minimal 2>&1 | artifact_out
 reset_fluentd_daemonset
-os::cmd::expect_success flush_fluentd_pos_files
-sudo rm -f /var/log/fluentd/fluentd.log
-oc label node --all logging-infra-fluentd=true 2>&1 | artifact_out
-os::cmd::try_until_text "oc get pods -l component=fluentd" "^logging-fluentd-.* Running "
+start_fluentd true 2>&1 | artifact_out
 fpod=$( get_running_pod fluentd )
 wait_for_fluentd_to_catch_up
 
 # configure fluentd to use MUX_CLIENT_MODE=maximal - verify logs get through
 os::log::info configure fluentd to use MUX_CLIENT_MODE=maximal - verify logs get through
-oc label node --all logging-infra-fluentd- 2>&1 | artifact_out
-os::cmd::try_until_text "oc get daemonset logging-fluentd -o jsonpath='{ .status.numberReady }'" "0" $FLUENTD_WAIT_TIME
+stop_fluentd "$fpod" $FLUENTD_WAIT_TIME 2>&1 | artifact_out
 oc set env daemonset/logging-fluentd MUX_CLIENT_MODE=maximal 2>&1 | artifact_out
 reset_fluentd_daemonset
-os::cmd::expect_success flush_fluentd_pos_files
-sudo rm -f /var/log/fluentd/fluentd.log
-oc label node --all logging-infra-fluentd=true 2>&1 | artifact_out
-os::cmd::try_until_text "oc get pods -l component=fluentd" "^logging-fluentd-.* Running "
+start_fluentd true 2>&1 | artifact_out
 fpod=$( get_running_pod fluentd )
 wait_for_fluentd_to_catch_up
